backend/requirements.txt
fastapi
uvicorn
openai
langgraph
chromadb
pydantic
requests
python-dotenv
